v1.36-20150213

+++ src/infinity/resource/key/BIFFResourceEntry.java
@@ -5,11 +5,20 @@
 package infinity.resource.key;
 
 import infinity.gui.BrowserMenuBar;
+import infinity.resource.Profile;
 import infinity.resource.ResourceFactory;
 import infinity.resource.Writeable;
-import infinity.util.*;
+import infinity.util.DynamicArray;
+import infinity.util.io.FileInputStreamNI;
+import infinity.util.io.FileNI;
+import infinity.util.io.FileReaderNI;
+import infinity.util.io.FileWriterNI;
 
-import java.io.*;
+import java.io.BufferedInputStream;
+import java.io.File;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
 
 public final class BIFFResourceEntry extends ResourceEntry implements Writeable
 {
@@ -32,23 +41,25 @@ public final class BIFFResourceEntry extends ResourceEntry implements Writeable
 
   public BIFFResourceEntry(byte buffer[], int offset, int stringLength)
   {
-    StringBuffer sb = new StringBuffer(Byteconvert.convertString(buffer, offset, stringLength));
-    type = (int)Byteconvert.convertShort(buffer, offset + stringLength);
-    locator = Byteconvert.convertInt(buffer, offset + stringLength + 2);
+    StringBuffer sb = new StringBuffer(DynamicArray.getString(buffer, offset, stringLength));
+    type = (int)DynamicArray.getShort(buffer, offset + stringLength);
+    locator = DynamicArray.getInt(buffer, offset + stringLength + 2);
     resourceName = sb.append('.').append(getExtension()).toString();
   }
 
 // --------------------- Begin Interface Writeable ---------------------
 
+  @Override
   public void write(OutputStream os) throws IOException
   {
-    Filewriter.writeString(os, resourceName.substring(0, resourceName.lastIndexOf((int)'.')), 8);
-    Filewriter.writeShort(os, (short)type);
-    Filewriter.writeInt(os, locator);
+    FileWriterNI.writeString(os, resourceName.substring(0, resourceName.lastIndexOf((int)'.')), 8);
+    FileWriterNI.writeShort(os, (short)type);
+    FileWriterNI.writeInt(os, locator);
   }
 
 // --------------------- End Interface Writeable ---------------------
 
+  @Override
   public boolean equals(Object o)
   {
     if (!(o instanceof BIFFResourceEntry))
@@ -57,6 +68,7 @@ public final class BIFFResourceEntry extends ResourceEntry implements Writeable
     return locator == other.locator && resourceName.equals(other.resourceName) && type == other.type;
   }
 
+  @Override
   public String toString()
   {
     return resourceName;
@@ -64,19 +76,20 @@ public final class BIFFResourceEntry extends ResourceEntry implements Writeable
 
   public void deleteOverride()
   {
-    File override = new File(ResourceFactory.getRootDir(),
-                             ResourceFactory.OVERRIDEFOLDER + File.separatorChar + resourceName);
-    if (override != null && override.exists())
+    File override = FileNI.getFile(Profile.getRootFolders(),
+                             Profile.getOverrideFolderName() + File.separatorChar + resourceName);
+    if (override != null && override.exists() && !override.isDirectory())
       override.delete();
     hasOverride = false;
   }
 
+  @Override
   public File getActualFile(boolean ignoreoverride)
   {
     if (!ignoreoverride) {
-      File override = new File(ResourceFactory.getRootDir(),
-                               ResourceFactory.OVERRIDEFOLDER + File.separatorChar + resourceName);
-      if (override.exists())
+      File override = FileNI.getFile(Profile.getRootFolders(),
+                               Profile.getOverrideFolderName() + File.separatorChar + resourceName);
+      if (override.exists() && !override.isDirectory())
         return override;
     }
     try {
@@ -93,6 +106,7 @@ public final class BIFFResourceEntry extends ResourceEntry implements Writeable
     return ResourceFactory.getKeyfile().getBIFFEntry(sourceindex);
   }
 
+  @Override
   public String getExtension()
   {
     String ext = ResourceFactory.getKeyfile().getExtension(type);
@@ -106,14 +120,15 @@ public final class BIFFResourceEntry extends ResourceEntry implements Writeable
     return locator;
   }
 
+  @Override
   public byte[] getResourceData(boolean ignoreoverride) throws Exception
   {
     if (!ignoreoverride) {
-      File override = new File(ResourceFactory.getRootDir(),
-                               ResourceFactory.OVERRIDEFOLDER + File.separatorChar + resourceName);
-      if (override.exists()) {
-        InputStream is = new BufferedInputStream(new FileInputStream(override));
-        byte buffer[] = Filereader.readBytes(is, (int)override.length());
+      File override = FileNI.getFile(Profile.getRootFolders(),
+                               Profile.getOverrideFolderName() + File.separatorChar + resourceName);
+      if (override.exists() && !override.isDirectory()) {
+        InputStream is = new BufferedInputStream(new FileInputStreamNI(override));
+        byte buffer[] = FileReaderNI.readBytes(is, (int)override.length());
         is.close();
         return buffer;
       }
@@ -124,13 +139,14 @@ public final class BIFFResourceEntry extends ResourceEntry implements Writeable
     return biff.getResource(locator & 0x3fff, false);
   }
 
+  @Override
   public InputStream getResourceDataAsStream(boolean ignoreoverride) throws Exception
   {
     if (!ignoreoverride) {
-      File override = new File(ResourceFactory.getRootDir(),
-                               ResourceFactory.OVERRIDEFOLDER + File.separatorChar + resourceName);
-      if (override.exists()) {
-        return new BufferedInputStream(new FileInputStream(override));
+      File override = FileNI.getFile(Profile.getRootFolders(),
+                               Profile.getOverrideFolderName() + File.separatorChar + resourceName);
+      if (override.exists() && !override.isDirectory()) {
+        return new BufferedInputStream(new FileInputStreamNI(override));
       }
     }
     BIFFArchive biff = ResourceFactory.getKeyfile().getBIFFFile(getBIFFEntry());
@@ -139,12 +155,13 @@ public final class BIFFResourceEntry extends ResourceEntry implements Writeable
     return biff.getResourceAsStream(locator & 0x3fff, false);
   }
 
+  @Override
   public int[] getResourceInfo(boolean ignoreoverride) throws IOException
   {
     if (!ignoreoverride) {
-      File override = new File(ResourceFactory.getRootDir(),
-                               ResourceFactory.OVERRIDEFOLDER + File.separatorChar + resourceName);
-      if (override.exists())
+      File override = FileNI.getFile(Profile.getRootFolders(),
+                               Profile.getOverrideFolderName() + File.separatorChar + resourceName);
+      if (override.exists() && !override.isDirectory())
         return getLocalFileInfo(override);
     }
     BIFFArchive biff = ResourceFactory.getKeyfile().getBIFFFile(getBIFFEntry());
@@ -153,17 +170,20 @@ public final class BIFFResourceEntry extends ResourceEntry implements Writeable
     return biff.getResourceInfo(locator & 0x3fff, false);
   }
 
+  @Override
   public String getResourceName()
   {
     return resourceName;
   }
 
+  @Override
   public String getTreeFolder()
   {
     if (BrowserMenuBar.getInstance() != null &&
         BrowserMenuBar.getInstance().getOverrideMode() == BrowserMenuBar.OVERRIDE_IN_OVERRIDE &&
-        hasOverride())
-      return ResourceFactory.OVERRIDEFOLDER;
+        hasOverride()) {
+      return Profile.getOverrideFolderName();
+    }
     return getExtension();
   }
 
@@ -172,11 +192,14 @@ public final class BIFFResourceEntry extends ResourceEntry implements Writeable
     return type;
   }
 
+  @Override
   public boolean hasOverride()
   {
-    if (!BrowserMenuBar.getInstance().cacheOverride())
-      hasOverride = new File(ResourceFactory.getRootDir(),
-                             ResourceFactory.OVERRIDEFOLDER + File.separatorChar + resourceName).exists();
+    if (!BrowserMenuBar.getInstance().cacheOverride()) {
+      File override = FileNI.getFile(Profile.getRootFolders(),
+                               Profile.getOverrideFolderName() + File.separatorChar + resourceName);
+      hasOverride = override.exists() && !override.isDirectory();
+    }
     return hasOverride;
   }
